-  update AgentDaoTest.kta -> add closeDb method at the end og each test

+++ app/src/main/java/com/openclassrooms/realestatemanager/db/dao/AgentDao.kt
@@ -17,7 +17,7 @@ interface AgentDao {
 
     //Get AgentEntity by Id
     @Query("SELECT * FROM agent WHERE id = :id")
-    fun getAgentById(id: Int): Flow<AgentEntity>
+    fun getAgentById(id: Long): Flow<AgentEntity>
 
     //Update AgentEntity
     @Update
